[GL] Move AsciiRasterInterface to GeoLib/IO.
