+++ FileIO/CMakeLists.txt
@@ -1,6 +1,4 @@
 set(SOURCES
-	AsciiRasterInterface.h
-	AsciiRasterInterface.cpp
 	CsvInterface.h
 	CsvInterface.cpp
 	GMSInterface.h
